requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib
python-jose
requests
beautifulsoup4
langchain-community
groq
python-multipart
bcrypt
mysql-connector-python
pymysql
pydantic[email]
oauthlib
stripe
jinja2
python-dotenv